+++ src/sync/state.rs
@@ -1,20 +1,3 @@
-// This file is part of Neverest, a CLI to synchronize emails.
-//
-// Copyright (C) 2024-2026  soywod <pimalaya.org@posteo.net>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
 //! Per-account JSON snapshot persisting, per side and mailbox, the
 //! content-keyed message set and the LCD checkpoints.
 
@@ -26,7 +9,7 @@ use std::{
 };
 
 use anyhow::{Context, Result, bail};
-use io_email::flag::Flag;
+use io_email::flag::types::Flag;
 use serde::{Deserialize, Serialize};
 
 use crate::{